+++ mitmproxy/addons/termlog.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 import asyncio
 import logging
 from typing import IO